Support for the cross-platfrom build on MacOS platform (HOST=x86_64-apple-darwin14)
  -changed min MacOS version to 10.14
  -set native toolchain for cross-platform builds to clang
  -improved check for package sources and cached packages
  -disabled ELF hardening feature in build.sh (--disable-hardening) for clang darwin platform

updated package build info:
  -bdb
  -librustzcash: build for darwin using native_rust, updated build options/target files

added new package definitions to support darwin build:
  -native_cctools - Apple cctools port for Linux and *BSD
  -native_clang - cross-platform toolchain to support C, C++ for Apple
  -native_b2 - boost build tool
  -native_rust - rust package

upgraded packages:
  boost 1.66 -> 1.75
  googletest 1.8.0 -> 1.8.1, fixed build for darwin
  rust 1.28 -> 1.49 (native)

build system changes:
  -added support for depends packages debugging using FIXED_RECIPE_HASH, set this variable in funсs.mk to
  disable depend packages rebuild on any makefile changes (sets fixed build ids for the packages)
  -added %_exact_file_name, %__exact_sha256_hash, %_exact_download_file, %_exact_download_path defines to
  override defaults for the specific platforms

Pastel source code changes:
 -boost bind placeholders compatibility fixes
 -changed datatype for timestamp in CPastelTicketBase from time_t to int64_t (portable on all platforms)

+++ src/mnode-rpc.cpp
@@ -2399,7 +2399,7 @@ UniValue ingest(const UniValue& params, bool fHelp) {
             }
 
             UniValue obj(UniValue::VOBJ);
-            obj.push_back(Pair(wtx.GetHash().GetHex(), lines));
+            obj.push_back(Pair(wtx.GetHash().GetHex(), (uint64_t)lines));
             mnObj.push_back(Pair(std::to_string(txCounter), obj));
 
             outfile << wtx.GetHash().GetHex() << " : " << lineCounter+1 << "-" << lineCounter+lines << " (" << lines << ")\n";
